docker-tg-bot/requirements.txt
python-telegram-bot[asyncio]>=21.0.0
aiodocker>=0.23.0
pytest>=8.0.0
pytest-asyncio>=0.24.0

